red: added StaticTypingFeature
